requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==8.0.3
colorama==0.4.4
decorator==5.1.0
feedparser==6.0.8
flasgger==0.9.5
Flask==2.0.2
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.3.1
lxml==4.7.1
MarkupSafe==2.0.1
mistune==2.0.0
PyJWT==2.3.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
sgmllib3k==1.0.0
six==1.16.0
SQLAlchemy==1.4.28
validators==0.18.2
Werkzeug==2.0.2
